[TEST] Add TimeTestDevice, tc_68_10. Use int32
- Add TimeTestDevice to runtime_util and Makefile
- Add tc_68_10: Latency test
- Convert int16_t to int32_t in runtime_util INT
- Add action_interval param to lowcar_protocol()
- Remove hardcoding type in virtual device code

+++ runtime_util/runtime_util.h
@@ -62,7 +62,7 @@ typedef enum param_type {
 
 //hold a single param value. One-to-one mapping to param_val_t enum
 typedef union {
-	int16_t p_i; //data if int
+	int32_t p_i; //data if int
 	float p_f;   //data if float
 	uint8_t p_b; //data if bool
 } param_val_t;
